moved the contact us routes to the routes folder
